Add task and update ui accordingly
